// pocketbase-types.ts
export enum Collections {
	Authorigins = "_authOrigins",
	Externalauths = "_externalAuths",
	Mfas = "_mfas",
	Otps = "_otps",
	Superusers = "_superusers",
	Folders = "folders",
	Notes = "notes",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type RecordIdString = string
export type HTMLString = string

// System fields
export type BaseSystemFields<T = never> = {
	id: RecordIdString
	collectionId: string
	collectionName: Collections
	expand?: T
}
// ...
export type FoldersRecord = {
	created?: IsoDateString
	id: string
	metadata?: string
	name: string
	updated?: IsoDateString
	user: RecordIdString
}

export type NotesRecord = {
	content?: string
	created?: IsoDateString
	folder?: RecordIdString
	id: string
	title: string
	updated?: IsoDateString
	user: RecordIdString
}
// ...
export type FoldersResponse<Texpand = unknown> = Required<FoldersRecord> & BaseSystemFields<Texpand>
export type NotesResponse<Texpand = unknown> = Required<NotesRecord> & BaseSystemFields<Texpand>